New PackageManagerViewModel
- introduced a single PackageManagerViewModel containing all view models necessary for the separate functionalities of the PacakgeManager
